+++ op-node/rollup/driver/metered_l1fetcher.go
@@ -57,6 +57,11 @@ func (m *MeteredL1Fetcher) FetchReceipts(ctx context.Context, blockHash common.H
 	return m.inner.FetchReceipts(ctx, blockHash)
 }
 
+func (m *MeteredL1Fetcher) PreFetchReceipts(ctx context.Context, blockHash common.Hash) (bool, error) {
+	defer m.recordTime("PreFetchReceipts")()
+	return m.inner.PreFetchReceipts(ctx, blockHash)
+}
+
 var _ derive.L1Fetcher = (*MeteredL1Fetcher)(nil)
 
 func (m *MeteredL1Fetcher) recordTime(method string) func() {
@@ -70,4 +75,9 @@ func (m *MeteredL1Fetcher) recordTime(method string) func() {
 func (m *MeteredL1Fetcher) GoOrUpdatePreFetchReceipts(ctx context.Context, l1StartBlock uint64) error {
 	defer m.recordTime("GoOrUpdatePreFetchReceipts")()
 	return m.inner.GoOrUpdatePreFetchReceipts(ctx, l1StartBlock)
+}
+
+func (m *MeteredL1Fetcher) ClearReceiptsCacheBefore(blockNumber uint64) {
+	defer m.recordTime("ClearReceiptsCacheBefore")()
+	m.inner.ClearReceiptsCacheBefore(blockNumber)
 }