+++ op-node/rollup/driver/sequencer.go
@@ -15,9 +15,16 @@ import (
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 )
 
+// When block produce is interrupted by high L1 latency, sequencer will build a full block periodically to avoid chain stuck
+const buildFullBlockInterval = 20
+
+// When block produce is lagging exceed lagTimeWindow, sequencer will set attrs.NoTxPool to true to quickly catch up
+const lagTimeWindow = 2 * time.Minute
+
 type Downloader interface {
 	InfoByHash(ctx context.Context, hash common.Hash) (eth.BlockInfo, error)
 	FetchReceipts(ctx context.Context, blockHash common.Hash) (eth.BlockInfo, types.Receipts, error)
+	PreFetchReceipts(ctx context.Context, blockHash common.Hash) (bool, error)
 }
 
 type L1OriginSelectorIface interface {
@@ -45,6 +52,9 @@ type Sequencer struct {
 	timeNow func() time.Time
 
 	nextAction time.Time
+
+	// if accEmptyBlocks > buildFullBlockInterval, will delay nextAction 600ms for full block building
+	accEmptyBlocks int
 }
 
 func NewSequencer(log log.Logger, cfg *rollup.Config, engine derive.ResettableEngineControl, attributesBuilder derive.AttributesBuilder, l1OriginSelector L1OriginSelectorIface, metrics SequencerMetrics) *Sequencer {
@@ -89,7 +99,19 @@ func (d *Sequencer) StartBuildingBlock(ctx context.Context) error {
 	// empty blocks (other than the L1 info deposit and any user deposits). We handle this by
 	// setting NoTxPool to true, which will cause the Sequencer to not include any transactions
 	// from the transaction pool.
-	attrs.NoTxPool = uint64(attrs.Timestamp) > l1Origin.Time+d.config.MaxSequencerDrift
+	if uint64(attrs.Timestamp) > l1Origin.Time+d.config.MaxSequencerDrift {
+		attrs.NoTxPool = true
+	} else {
+		// This is short term solution to increase sequencer catching up speed.
+		// Long term solution should optimize op-geth payload building work flow.
+		attrsTime := time.Unix(int64(attrs.Timestamp), 0)
+		isCatchingUp := time.Since(attrsTime) > lagTimeWindow
+		if isCatchingUp && (d.accEmptyBlocks < buildFullBlockInterval) {
+			attrs.NoTxPool = true
+		} else {
+			attrs.NoTxPool = false
+		}
+	}
 
 	d.log.Debug("prepared attributes for new block",
 		"num", l2Head.Number+1, "time", uint64(attrs.Timestamp),
@@ -227,6 +249,9 @@ func (d *Sequencer) RunNextSequencerAction(ctx context.Context) (*eth.ExecutionP
 			return nil, nil
 		} else {
 			d.attrBuilder.CachePayloadByHash(payload)
+			if len(payload.Transactions) == 1 {
+				d.accEmptyBlocks += 1
+			}
 			d.log.Info("sequencer successfully built a new block", "block", payload.ID(), "time", uint64(payload.Timestamp), "txs", len(payload.Transactions))
 			return payload, nil
 		}
@@ -249,6 +274,11 @@ func (d *Sequencer) RunNextSequencerAction(ctx context.Context) (*eth.ExecutionP
 			}
 		} else {
 			parent, buildingID, _ := d.engine.BuildingPayload() // we should have a new payload ID now that we're building a block
+			if d.accEmptyBlocks >= buildFullBlockInterval {
+				d.nextAction = d.timeNow().Add(600 * time.Millisecond)
+				d.accEmptyBlocks = 0
+				d.log.Info("sequencer delay next action 600ms and reset accEmptyBlocks")
+			}
 			d.log.Info("sequencer started building new block", "payload_id", buildingID, "l2_parent_block", parent, "l2_parent_block_time", parent.Time)
 		}
 		return nil, nil