+++ op-service/client/client.go
@@ -52,8 +52,12 @@ type Client interface {
 // InstrumentedClient is an Ethereum client that tracks
 // Prometheus metrics for each call.
 type InstrumentedClient struct {
-	c *ethclient.Client
-	m *metrics.Metrics
+	c Client
+	m Metricer
+}
+
+type Metricer interface {
+	RecordRPCClientRequest(method string) func(err error)
 }
 
 // NewInstrumentedClient creates a new instrumented client. It takes
@@ -62,6 +66,13 @@ type InstrumentedClient struct {
 func NewInstrumentedClient(c *rpc.Client, m *metrics.Metrics) *InstrumentedClient {
 	return &InstrumentedClient{
 		c: ethclient.NewClient(c),
+		m: m,
+	}
+}
+
+func NewInstrumentedClientWithClient(c Client, m Metricer) *InstrumentedClient {
+	return &InstrumentedClient{
+		c: c,
 		m: m,
 	}
 }
@@ -263,14 +274,14 @@ func (ic *InstrumentedClient) SendTransaction(ctx context.Context, tx *types.Tra
 	})
 }
 
-func instrument1(m *metrics.Metrics, name string, cb func() error) error {
+func instrument1(m Metricer, name string, cb func() error) error {
 	record := m.RecordRPCClientRequest(name)
 	err := cb()
 	record(err)
 	return err
 }
 
-func instrument2[O any](m *metrics.Metrics, name string, cb func() (O, error)) (O, error) {
+func instrument2[O any](m Metricer, name string, cb func() (O, error)) (O, error) {
 	record := m.RecordRPCClientRequest(name)
 	res, err := cb()
 	record(err)