+++ op-service/client/fallback_client.go
@@ -154,6 +154,22 @@ func (l *FallbackClient) CallContract(ctx context.Context, call ethereum.CallMsg
 	return contract, err
 }
 
+func (l *FallbackClient) BlockByNumber(ctx context.Context, number *big.Int) (*types.Block, error) {
+	block, err := (*l.currentClient.Load()).BlockByNumber(ctx, number)
+	if err != nil {
+		l.handleErr(err, "BlockByNumber")
+	}
+	return block, err
+}
+
+func (l *FallbackClient) BlockByHash(ctx context.Context, hash common.Hash) (*types.Block, error) {
+	block, err := (*l.currentClient.Load()).BlockByHash(ctx, hash)
+	if err != nil {
+		l.handleErr(err, "BlockByHash")
+	}
+	return block, err
+}
+
 func (l *FallbackClient) Close() {
 	l.mx.Lock()
 	defer l.mx.Unlock()