+++ op-service/txmgr/metrics/noop.go
@@ -12,3 +12,7 @@ func (*NoopTxMetrics) TxConfirmed(*types.Receipt)        {}
 func (*NoopTxMetrics) TxPublished(string)                {}
 func (*NoopTxMetrics) RPCError()                         {}
 func (m *NoopTxMetrics) RecordL1UrlSwitchEvt(url string) {}
+func (m *NoopTxMetrics) RecordRPCClientRequest(method string) func(err error) {
+	return func(err error) {
+	}
+}