+++ op-service/txmgr/metrics/tx_metrics.go
@@ -1,10 +1,15 @@
 package metrics
 
 import (
+	"errors"
+	"fmt"
+
 	"github.com/ethereum-optimism/optimism/op-service/client"
 	"github.com/ethereum-optimism/optimism/op-service/metrics"
+	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/params"
+	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/prometheus/client_golang/prometheus"
 )
@@ -18,20 +23,24 @@ type TxMetricer interface {
 	TxPublished(string)
 	RPCError()
 	client.FallbackClientMetricer
+	client.Metricer
 }
 
 type TxMetrics struct {
-	TxL1GasFee         prometheus.Gauge
-	txFees             prometheus.Counter
-	TxGasBump          prometheus.Gauge
-	txFeeHistogram     prometheus.Histogram
-	LatencyConfirmedTx prometheus.Gauge
-	currentNonce       prometheus.Gauge
-	pendingTxs         prometheus.Gauge
-	txPublishError     *prometheus.CounterVec
-	publishEvent       *metrics.Event
-	confirmEvent       metrics.EventVec
-	rpcError           prometheus.Counter
+	TxL1GasFee                      prometheus.Gauge
+	txFees                          prometheus.Counter
+	TxGasBump                       prometheus.Gauge
+	txFeeHistogram                  prometheus.Histogram
+	LatencyConfirmedTx              prometheus.Gauge
+	currentNonce                    prometheus.Gauge
+	pendingTxs                      prometheus.Gauge
+	txPublishError                  *prometheus.CounterVec
+	publishEvent                    *metrics.Event
+	confirmEvent                    metrics.EventVec
+	rpcError                        prometheus.Counter
+	RPCClientRequestsTotal          *prometheus.CounterVec
+	RPCClientRequestDurationSeconds *prometheus.HistogramVec
+	RPCClientResponsesTotal         *prometheus.CounterVec
 	*client.FallbackClientMetrics
 }
 
@@ -108,6 +117,32 @@ func MakeTxMetrics(ns string, factory metrics.Factory) TxMetrics {
 			Subsystem: "txmgr",
 		}),
 		FallbackClientMetrics: client.NewFallbackClientMetrics(ns, factory),
+		RPCClientRequestsTotal: factory.NewCounterVec(prometheus.CounterOpts{
+			Namespace: ns,
+			Subsystem: "txmgr_rpc_client",
+			Name:      "requests_total",
+			Help:      "Total RPC requests initiated by the txmgr's RPC client",
+		}, []string{
+			"method",
+		}),
+		RPCClientRequestDurationSeconds: factory.NewHistogramVec(prometheus.HistogramOpts{
+			Namespace: ns,
+			Subsystem: "txmgr_rpc_client",
+			Name:      "request_duration_seconds",
+			Buckets:   []float64{.005, .01, .025, .05, .1, .25, .5, 1, 2.5, 5, 10},
+			Help:      "Histogram of RPC client request durations",
+		}, []string{
+			"method",
+		}),
+		RPCClientResponsesTotal: factory.NewCounterVec(prometheus.CounterOpts{
+			Namespace: ns,
+			Subsystem: "txmgr_rpc_client",
+			Name:      "responses_total",
+			Help:      "Total RPC request responses received by the txmgr's RPC client",
+		}, []string{
+			"method",
+			"error",
+		}),
 	}
 }
 
@@ -147,4 +182,37 @@ func (t *TxMetrics) TxPublished(errString string) {
 
 func (t *TxMetrics) RPCError() {
 	t.rpcError.Inc()
+}
+
+func (t *TxMetrics) RecordRPCClientRequest(method string) func(err error) {
+	t.RPCClientRequestsTotal.WithLabelValues(method).Inc()
+	timer := prometheus.NewTimer(t.RPCClientRequestDurationSeconds.WithLabelValues(method))
+	return func(err error) {
+		t.RecordRPCClientResponse(method, err)
+		timer.ObserveDuration()
+	}
+}
+
+// RecordRPCClientResponse records an RPC response. It will
+// convert the passed-in error into something metrics friendly.
+// Nil errors get converted into <nil>, RPC errors are converted
+// into rpc_<error code>, HTTP errors are converted into
+// http_<status code>, and everything else is converted into
+// <unknown>.
+func (t *TxMetrics) RecordRPCClientResponse(method string, err error) {
+	var errStr string
+	var rpcErr rpc.Error
+	var httpErr rpc.HTTPError
+	if err == nil {
+		errStr = "<nil>"
+	} else if errors.As(err, &rpcErr) {
+		errStr = fmt.Sprintf("rpc_%d", rpcErr.ErrorCode())
+	} else if errors.As(err, &httpErr) {
+		errStr = fmt.Sprintf("http_%d", httpErr.StatusCode)
+	} else if errors.Is(err, ethereum.NotFound) {
+		errStr = "<not found>"
+	} else {
+		errStr = "<unknown>"
+	}
+	t.RPCClientResponsesTotal.WithLabelValues(method, errStr).Inc()
 }